Revert "Refactored Table to reduce files"
This reverts commit 714d9a3068d49b1ad0dad5293275befc4ba1d734.
